_import/pioreactorui/requirements.txt
# pyyaml is installed elsewhere
# this should be a superset of requirements_leader.txt
flask==3.1.0
flup6==1.1.1
python-dotenv==1.0.1
paho-mqtt==2.1.0
huey==2.5.0
msgspec==0.19.0
werkzeug==3.1.3
pioreactor>=24.9.26
